Added multiple files load Refactoring
