+++ src/main/java/technikal/task/fishmarket/controllers/FishController.java
@@ -8,7 +8,7 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import technikal.task.fishmarket.entity.Fish;
-import technikal.task.fishmarket.models.FishDto;
+import technikal.task.fishmarket.dto.FishDto;
 import technikal.task.fishmarket.services.FishService;
 
 import java.util.List;
