+++ src/main/java/technikal/task/fishmarket/entity/Fish.java
@@ -2,14 +2,17 @@ package technikal.task.fishmarket.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
 @Entity
 @Getter
 @Setter
+@NoArgsConstructor
 @Table(name = "fish")
 public class Fish {
 
@@ -19,11 +22,18 @@ public class Fish {
 	private int id;
 	
 	private String name;
-	private double price;
-	private Date catchDate;
-//	private String imageFileName;
 
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "fish", cascade = CascadeType.ALL)
-	private List<DataFile> attachedFiles;
+	private double price;
+
+	@Column(name = "catch_date")
+	private Date catchDate;
+
+	@OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL, mappedBy = "fish")
+	private List<DataFile> attachedFiles = new ArrayList<>();
+
+	public void addFile(DataFile file) {
+		file.setFish(this);
+		this.attachedFiles.add(file);
+	}
 
 }
