+++ src/main/java/technikal/task/fishmarket/repository/FishRepository.java
@@ -2,6 +2,8 @@ package technikal.task.fishmarket.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import org.springframework.data.jpa.repository.Query;
 import technikal.task.fishmarket.entity.Fish;
 
-public interface FishRepository extends JpaRepository<Fish, Integer> {}
+public interface FishRepository extends JpaRepository<Fish, Integer> {
+}
