+++ src/main/java/technikal/task/fishmarket/services/FishService.java
@@ -1,19 +1,18 @@
 package technikal.task.fishmarket.services;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
+import technikal.task.fishmarket.dto.FishDto;
 import technikal.task.fishmarket.entity.DataFile;
 import technikal.task.fishmarket.entity.Fish;
 import technikal.task.fishmarket.exception.exceptions.FishCreationException;
-import technikal.task.fishmarket.models.FishDto;
-import technikal.task.fishmarket.repository.DataFilesRepository;
 import technikal.task.fishmarket.repository.FishRepository;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
@@ -28,7 +27,6 @@ import java.util.List;
 public class FishService {
 
     private final FishRepository fishRepository;
-    private final DataFilesRepository dataFilesRepository;
     private final StorageService storageService;
 
     private final Sort defaultValue = Sort.by(Sort.Direction.DESC, "id");
@@ -37,6 +35,7 @@ public class FishService {
         return fishRepository.findAll(defaultValue);
     }
 
+    @Transactional
     public void addFish(FishDto fishDto, BindingResult result) {
         if (fishDto == null) throw new NullPointerException("Provided fishDto is null");
 
@@ -54,21 +53,19 @@ public class FishService {
         fish.setCatchDate(date);
         fish.setName(fishDto.getName());
         fish.setPrice(fishDto.getPrice());
-        fish.setAttachedFiles(new ArrayList<>());
 
-        List<DataFile> dataFiles = Arrays.stream(fishDto.getImageFiles())
+        List<DataFile> files = Arrays.stream(fishDto.getImageFiles())
                 .map(data -> {
                     DataFile dataFile = new DataFile();
-                    dataFile.setFish(fish);
                     dataFile.setFileType(data.getContentType());
                     dataFile.setSaveDate(date);
                     dataFile.setFileName(date.getTime() + "_" + data.getOriginalFilename());
-                    fish.getAttachedFiles().add(dataFile);
+                    dataFile.setFish(fish);
                     return dataFile;
                 })
                 .toList();
+        fish.setAttachedFiles(files);
         fishRepository.save(fish);
-        dataFilesRepository.saveAll(dataFiles);
         storageService.saveImages(fishDto.getImageFiles(), date);
     }
