+++ src/main/resources/application-dev.properties
@@ -2,13 +2,16 @@ server.port=8081
 
 spring.application.name=fishmarket
 
-#spring.datasource.url=jdbc:h2:mem:testdb for in-memory
+# for in-memory
+spring.datasource.url=jdbc:h2:mem:testdb;MODE=LEGACY
 
-spring.datasource.url=jdbc:h2:file:/data/demo
+#spring.datasource.url=jdbc:h2:file:/data/demo;MODE=LEGACY
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+
+spring.h2.console.enabled=true
 
 # Spring Security
 #spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration
