+++ src/main/resources/templates/createFish.html
@@ -42,7 +42,8 @@
 						<lebel class="col-sm-4 col-form-label">Фото</lebel>
 						<div class="col-sm-8">
 							<input class="form-control" type="file" step="0.01" min="0"
-								th:field="${fishDto.imageFiles}" multiple>
+								th:field="${fishDto.imageFiles}" multiple onchange="validateFileCount(this, 3)">
+							<p id="file-error" class="text-danger" style="display: none;">Дозволено максимум 3 фото</p>
 							<p th:if="${#fields.hasErrors('imageFiles')}"
 								th:errorclass="text-danger" th:errors="${fishDto.imageFiles}"></p>
 						</div>
@@ -66,6 +67,19 @@
 	<script
 		src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
 		integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
-		crossorigin="anonymous"></script>
+		crossorigin="anonymous">
+		function validateFileCount(input, maxFiles) {
+			const fileCount = input.files.length;
+			const errorElement = document.getElementById('file-error');
+
+			if (fileCount > maxFiles) {
+				errorElement.style.display = 'block';
+				input.value = ''; // Clear the input if the limit is exceeded
+			} else {
+				errorElement.style.display = 'none';
+			}
+		}
+
+	</script>
 </body>
 </html>
