+++ src/main/resources/templates/index.html
@@ -50,7 +50,7 @@
               <div class="image-container">
 <!--                <img th:src="@{'/images/'+${fish.imageFileName}}" alt="..." width="100">-->
                 <div th:each="data : ${fish.attachedFiles}">
-                  <img th:src="@{'/images/'+${data.fileName}}" alt="..." width="50"/>
+                  <img th:src="@{'public/images/'+${data.fileName}}" alt="..." width="50"/>
                 </div>
               </div>
             </td>
